charts appears and loads data from core data, does not reload chart after new intake

+++ Calorie Tracker/Calorie Tracker/Model Controller/CalorieIntakeController.swift
@@ -19,7 +19,7 @@ class CalorieIntakeController {
         }
     }
     
-    func create(calories: Int16) {
+    func create(calories: Double) {
         
         _ = CalorieIntake(calories: calories)
         saveToPersistentStore()
